+++ .prettierrc
@@ -1,7 +1,7 @@
 {
-    "printWidth": 100,
-    "singleQuote": true,
-    "tabWidth": 4,
-    "arrowParens": "always",
-    "trailingComma": "es5"
+  "printWidth": 100,
+  "singleQuote": true,
+  "tabWidth": 4,
+  "arrowParens": "always",
+  "trailingComma": "es5"
 }
