+++ index.html
@@ -1,17 +1,17 @@
 <!doctype html>
 <html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-        <meta name="theme-color" content="#000000" />
-        <link rel="manifest" href="manifest.json" />
-        <link rel="shortcut icon" href="favicon.ico" />
-        <title>Riverflow : react version</title>
-        <meta name="description" content="get current Texas river flow cfs from the USGS api" />
-    </head>
-    <body>
-        <noscript>You need to enable JavaScript to run this app. Get with the times man.</noscript>
-        <div id="root"></div>
-        <script type="module" src="/src/index.js"></script>
-    </body>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+    <meta name="theme-color" content="#000000" />
+    <link rel="manifest" href="manifest.json" />
+    <link rel="shortcut icon" href="favicon.ico" />
+    <title>Riverflow : react version</title>
+    <meta name="description" content="get current Texas river flow cfs from the USGS api" />
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app. Get with the times man.</noscript>
+    <div id="root"></div>
+    <script type="module" src="/src/index.jsx"></script>
+  </body>
 </html>
