+++ package.json
@@ -1,35 +1,38 @@
 {
-    "name": "react-riverflow",
-    "homepage": "https://mountaindrawn.com/react-riverflow",
-    "version": "0.1.0",
-    "private": true,
-    "type": "module",
-    "scripts": {
-        "dev": "vite",
-        "start": "vite",
-        "build": "vite build",
-        "preview": "vite preview",
-        "test": "vitest run",
-        "test:watch": "vitest",
-        "deploy": "npm run build && gh-pages -d dist"
-    },
-    "dependencies": {
-        "bulma": "^0.7.2",
-        "classnames": "^2.5.1",
-        "prop-types": "^15.8.1",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
-        "react-router-dom": "^6.30.0"
-    },
-    "devDependencies": {
-        "@testing-library/dom": "^10.4.0",
-        "@testing-library/jest-dom": "^6.6.3",
-        "@testing-library/react": "^16.1.0",
-        "@vitejs/plugin-react": "^4.3.4",
-        "gh-pages": "^6.2.0",
-        "jsdom": "^25.0.1",
-        "sass": "^1.83.0",
-        "vite": "^5.4.11",
-        "vitest": "^2.1.8"
-    }
+  "name": "react-riverflow",
+  "version": "0.1.0",
+  "private": true,
+  "homepage": "https://mountaindrawn.com/react-riverflow",
+  "type": "module",
+  "scripts": {
+    "dev": "vp dev",
+    "start": "vp dev",
+    "build": "vp build",
+    "preview": "vp preview",
+    "test": "vp test run",
+    "test:watch": "vp test",
+    "deploy": "vp build && gh-pages -d dist",
+    "prepare": "vp config"
+  },
+  "dependencies": {
+    "bulma": "^0.9.4",
+    "classnames": "^2.5.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.30.0"
+  },
+  "devDependencies": {
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@vitejs/plugin-react": "^6.0.2",
+    "gh-pages": "^6.2.0",
+    "jsdom": "^25.0.1",
+    "sass": "^1.83.0",
+    "vite": "8.0.16",
+    "vite-plus": "latest",
+    "vitest": "4.1.8"
+  },
+  "packageManager": "pnpm@10.33.4"
 }
