+++ src/components/Riverflow/Riverflow.jsx
@@ -1,11 +1,11 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import Classnames from 'classnames';
-import Gridtable from '../../components/Gridtable/Gridtable';
-import Rivers from '../../rivers.json';
-import Conditions from '../../conditions.json';
-import { Errors, Notification } from './components';
-import './Riverflow.sass';
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import Classnames from "classnames";
+import Gridtable from "../../components/Gridtable/Gridtable";
+import Rivers from "../../rivers.json";
+import Conditions from "../../conditions.json";
+import { Errors, Notification } from "./components";
+import "./Riverflow.sass";
 
 class Riverflow extends Component {
   static propTypes = {
@@ -21,14 +21,14 @@ class Riverflow extends Component {
 
     this.state = {
       error: false,
-      graphType: '00060', // defaults to cfs
+      graphType: "00060", // defaults to cfs
       loading: true,
-      searchQuery: '', // search filter
+      searchQuery: "", // search filter
       tableData: [],
     };
 
-    this.baseMapUrl = '//maps.google.com/?q=';
-    this.baseUsgsUrl = 'https://waterservices.usgs.gov/nwis/iv/';
+    this.baseMapUrl = "//maps.google.com/?q=";
+    this.baseUsgsUrl = "https://waterservices.usgs.gov/nwis/iv/";
     this.rivers = Rivers.data;
     this.riversFormatted = [];
     this.sites = this.formatSites();
@@ -46,12 +46,12 @@ class Riverflow extends Component {
     this.getUsgsData();
   }
 
-  handleRefreshTable = e => {
+  handleRefreshTable = (e) => {
     e.preventDefault();
     this.getUsgsData();
   };
 
-  handleFilterTable = e => {
+  handleFilterTable = (e) => {
     this.setState({
       searchQuery: e.target.value,
     });
@@ -59,12 +59,12 @@ class Riverflow extends Component {
 
   hideNotification(e) {
     e.preventDefault();
-    e.currentTarget.parentElement.classList.add('is-hidden');
+    e.currentTarget.parentElement.classList.add("is-hidden");
   }
 
-  clearSearch = e => {
+  clearSearch = (e) => {
     this.setState({
-      searchQuery: '',
+      searchQuery: "",
     });
   };
 
@@ -72,12 +72,12 @@ class Riverflow extends Component {
     const { error, graphType, loading, searchQuery, tableData } = this.state;
     const { selected } = this.props;
 
-    let refreshClasses = Classnames('button is-primary', {
-      'is-loading': this.state.loading,
+    let refreshClasses = Classnames("button is-primary", {
+      "is-loading": this.state.loading,
     });
 
-    let errorClasses = Classnames('notification is-danger', {
-      'is-hidden': this.state.error === false,
+    let errorClasses = Classnames("notification is-danger", {
+      "is-hidden": this.state.error === false,
     });
 
     return (
@@ -85,11 +85,7 @@ class Riverflow extends Component {
         <section className="section">
           <div className="container">
             <div className="Riverflow">
-              <Errors
-                classes={errorClasses}
-                error={error}
-                hide={this.hideNotification}
-              />
+              <Errors classes={errorClasses} error={error} hide={this.hideNotification} />
 
               <Notification hide={this.hideNotification} />
 
@@ -115,7 +111,7 @@ class Riverflow extends Component {
                         onClick={this.clearSearch}
                         aria-label="clear the search"
                       >
-                        {' '}
+                        {" "}
                       </button>
                     </p>
                   </div>
@@ -127,12 +123,8 @@ class Riverflow extends Component {
                     onClick={this.handleRefreshTable}
                     aria-label="Refresh the table data"
                   >
-                    <span className="refresh-long is-hidden-mobile">
-                      refresh river table
-                    </span>
-                    <span className="refresh-short is-hidden-tablet">
-                      &#8634;
-                    </span>
+                    <span className="refresh-long is-hidden-mobile">refresh river table</span>
+                    <span className="refresh-short is-hidden-tablet">&#8634;</span>
                   </button>
                 </div>
               </div>
@@ -155,9 +147,9 @@ class Riverflow extends Component {
 
   scrollIntoView() {
     window.setTimeout(() => {
-      const el = document.querySelector('.is-selected');
+      const el = document.querySelector(".is-selected");
       if (el) {
-        el.scrollIntoView({ behavior: 'smooth' });
+        el.scrollIntoView({ behavior: "smooth" });
       }
     }, 1000); // delay for usability
   }
@@ -165,7 +157,7 @@ class Riverflow extends Component {
   formatSites() {
     let list = [];
 
-    this.rivers.forEach(function(d) {
+    this.rivers.forEach(function (d) {
       // return only number values
       if (d.value.match(/\d+/g)) {
         list.push(d.value);
@@ -186,29 +178,29 @@ class Riverflow extends Component {
     const params = new URLSearchParams({
       parameterCd: this.state.graphType,
       sites: this.sites,
-      format: 'json',
-      period: 'PT12H', // past 12 hours
-      siteStatus: 'active',
+      format: "json",
+      period: "PT12H", // past 12 hours
+      siteStatus: "active",
     });
 
     fetch(`${this.baseUsgsUrl}?${params}`)
-      .then(response => {
+      .then((response) => {
         if (!response.ok) {
           throw new Error(`USGS request failed (${response.status})`);
         }
         return response.json();
       })
-      .then(data => {
+      .then((data) => {
         this.setState({ loading: false });
         if (data.value.timeSeries) {
           this.displayUsgsData(data.value.timeSeries);
           this.scrollIntoView();
           this.setState({ error: false });
         } else {
-          this.setState({ error: 'no river data available' });
+          this.setState({ error: "no river data available" });
         }
       })
-      .catch(error => {
+      .catch((error) => {
         console.log(error);
         this.setState({ loading: false, error: error.message });
       });
@@ -236,7 +228,7 @@ class Riverflow extends Component {
     let site;
     let time;
 
-    response.forEach(function(d, i, a) {
+    response.forEach(function (d, i, a) {
       // NOTE: some rivers do not support cfs (00060)
       arr = d.values[0].value;
       // return on error
@@ -250,8 +242,8 @@ class Riverflow extends Component {
       // get current date / time
       date = new Date(currentValue.dateTime);
       time = date.toLocaleTimeString([], {
-        hour: '2-digit',
-        minute: '2-digit',
+        hour: "2-digit",
+        minute: "2-digit",
         hour12: false,
       });
 
@@ -259,7 +251,7 @@ class Riverflow extends Component {
 
       // only show date if not today
       if (today.toDateString() === date.toDateString()) {
-        date = '';
+        date = "";
       }
 
       geo = d.sourceInfo.geoLocation.geogLocation;
@@ -269,7 +261,7 @@ class Riverflow extends Component {
 
       river = {
         name: d.sourceInfo.siteName,
-        location: that.baseMapUrl + geo.latitude + ',+' + geo.longitude,
+        location: that.baseMapUrl + geo.latitude + ",+" + geo.longitude,
         site: site,
         date: date,
         time: time,
@@ -292,7 +284,7 @@ class Riverflow extends Component {
    * @param {Object} river
    */
   mergeRiverInfo(river) {
-    this.rivers.forEach(function(d) {
+    this.rivers.forEach(function (d) {
       // add white water class
       if (d.value === river.site) {
         river.class = d.class;
@@ -315,25 +307,25 @@ class Riverflow extends Component {
     // check the range of the cfs and display the appropriate message
     if (cfs === 0) {
       condition = Conditions.flow0;
-      level = 'level-0';
+      level = "level-0";
     } else if (cfs > 0 && cfs < 50) {
       condition = Conditions.flow1;
-      level = 'level-1';
+      level = "level-1";
     } else if (cfs >= 50 && cfs < 100) {
       condition = Conditions.flow2;
-      level = 'level-2';
+      level = "level-2";
     } else if (cfs >= 100 && cfs < 300) {
       condition = Conditions.flow3;
-      level = 'level-3';
+      level = "level-3";
     } else if (cfs >= 300 && cfs < 600) {
       condition = Conditions.flow4;
-      level = 'level-4';
+      level = "level-4";
     } else if (cfs >= 600 && cfs < 2000) {
       condition = Conditions.flow5;
-      level = 'level-5';
+      level = "level-5";
     } else if (cfs >= 2000) {
       condition = Conditions.flow6;
-      level = 'level-6';
+      level = "level-6";
     }
 
     return { condition: condition, level: level };
