+++ src/conditions.json
@@ -1,9 +1,9 @@
 {
-	"flow0": "Sorry but this river is bone dry. Try a spring fed river like the San Marcos til we get more rain.",
-	"flow1": "The river is pretty much just a trickle right. Not much good for floating at the moment but a good rain should bring it up",
-	"flow2": "It's barely moving but it should be floatable in kayaks or tubes. Be prepared to drag bottom in spots though.",
-	"flow3": "This is a pretty leisurely flow but still fun. You shouldn't have any problems scraping bottom in canoes at this level",
-	"flow4": "Now we're talking! The river is flowing pretty good but not too dangerous. If the graph shows a sharp increase over the past week it may still be rising.",
-	"flow5": "The flow is moving now! Unless this is a large volume river like the Colorado or Rio Grande you can expect some really fast moving water.",
-	"flow6": "DANGER! This is flowing at a high volume! Unless you are an experienced paddler please check with a local outfitter for more information before heading out. If this is a large river like the Colorado it may be good this high."
+  "flow0": "Sorry but this river is bone dry. Try a spring fed river like the San Marcos til we get more rain.",
+  "flow1": "The river is pretty much just a trickle right. Not much good for floating at the moment but a good rain should bring it up",
+  "flow2": "It's barely moving but it should be floatable in kayaks or tubes. Be prepared to drag bottom in spots though.",
+  "flow3": "This is a pretty leisurely flow but still fun. You shouldn't have any problems scraping bottom in canoes at this level",
+  "flow4": "Now we're talking! The river is flowing pretty good but not too dangerous. If the graph shows a sharp increase over the past week it may still be rising.",
+  "flow5": "The flow is moving now! Unless this is a large volume river like the Colorado or Rio Grande you can expect some really fast moving water.",
+  "flow6": "DANGER! This is flowing at a high volume! Unless you are an experienced paddler please check with a local outfitter for more information before heading out. If this is a large river like the Colorado it may be good this high."
 }
