--- src/index.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import { HashRouter as Router, Routes, Route } from 'react-router-dom';
-import App from './App';
-import Error404 from './components/Error404';
-import './index.sass';
-
-const Routing = () => {
-	return (
-		<Router>
-			<Routes>
-				<Route path="/" element={<App />} />
-				<Route path="/:id" element={<App />} />
-				<Route path="*" element={<Error404 />} />
-			</Routes>
-		</Router>
-	);
-};
-
-const root = createRoot(document.getElementById('root'));
-root.render(<Routing />);
