+++ src/rivers.json
@@ -1,50 +1,52 @@
-{"data": [
-	{ "text": "Brazos River : Palo Pinto", "value": "08089000", "class": "II-III" },
-	{ "text": "Brazos River : Glen Rose", "value": "08091000", "class": "II-III" },
-	{ "text": "Brazos River : Waco", "value": "08096500", "class": "II-III" },
-	{ "text": "Brazos River : Bryan", "value": "08108700", "class": "II-III" },
-	{ "text": "Lampasas River : Kempner", "value": "08103800", "class": "II+" },
-	{ "text": "Lampasas River : Ding Dong", "value": "08103940", "class": "II+" },
-	{ "text": "Lampasas River : Belton", "value": "08104100", "class": "II+" },
-	{ "text": "Barton Springs", "value": "08155500", "class": "N/A" },
-	{ "text": "Barton Creek : Above Barton Springs", "value": "08155400", "class": "II" },
-	{ "text": "Barton Creek : Loop 360", "value": "08155300", "class": "II-III" },
-	{ "text": "Barton Creek : Lost Ck Blvd", "value": "08155240", "class": "II-III" },
-	{ "text": "Barton Creek : SH 71", "value": "08155200", "class": "II-III+" },
-	{ "text": "Colorado River : Austin", "value": "08158000", "class": "I-II" },
-	{ "text": "Colorado River : Bastrop", "value": "08159200", "class": "I-II" },
-	{ "text": "Onion Creek : Manchaca", "value": "08158827", "class": "II-III" },
-	{ "text": "Onion Creek : Hwy 183", "value": "08159000", "class": "II-III" },
-	{ "text": "Onion Creek : Driftwood", "value": "08158700", "class": "II-III" },
-	{ "text": "Llano River : Llano", "value": "08151500", "class": "II-III" },
-	{ "text": "Llano River : Mason", "value": "08150700", "class": "II-III" },
-	{ "text": "Llano River : Junction", "value": "08150000", "class": "I-II" },
-	{ "text": "San Saba River : Menard", "value": "08144500", "class": "II" },
-	{ "text": "San Saba River : San Saba", "value": "08146000", "class": "II" },
-	{ "text": "Guadalupe River : Spring Branch", "value": "08167500", "class": "II" },
-	{ "text": "Guadalupe River : Sattler", "value": "08167800", "class": "II" },
-	{ "text": "Guadalupe River : Comal New Braunfels", "value": "08168500", "class": "II" },
-	{ "text": "Guadalupe River : FM 1117 nr Seguin", "value": "08169792", "class": "II" },
-	{ "text": "Guadalupe River : Gonzales", "value": "08169845", "class": "II" },
-	{ "text": "San Marcos River : San Marcos", "value": "08170500", "class": "II" },
-	{ "text": "San Marcos River : Luling", "value": "08172000", "class": "II" },
-	{ "text": "Blanco River : Wimberley", "value": "08171000", "class": "II-III" },
-	{ "text": "Blanco River : Kyle", "value": "08171300", "class": "II-III" },
-	{ "text": "Nueces River : Montell", "value": "08189998", "class": "II" },
-	{ "text": "Nueces River : Laguna", "value": "08190000", "class": "II" },
-	{ "text": "Nueces River : Uvalde", "value": "08192000", "class": "II" },
-	{ "text": "Nueces River : Cotulla", "value": "08194000", "class": "II" },
-	{ "text": "Nueces River : Three Rivers", "value": "08210000", "class": "II" },
-	{ "text": "Frio River : Concan", "value": "08195000", "class": "II" },
-	{ "text": "Rio Grande : Castolon", "value": "08374550", "class": "II-III" },
-	{ "text": "Rio Grande : Big Bend", "value": "08375300", "class": "II-III" },
-	{ "text": "Pecos River : Orla", "value": "08412500", "class": "II-III" },
-	{ "text": "Pecos River : Girvin", "value": "08446500", "class": "II-III" },
-	{ "text": "Medina River : Patterson Road", "value": "0817887350", "class": "I-II" },
-	{ "text": "Medina River : Bandera", "value": "08178880", "class": "I-II" },
-	{ "text": "San Antonio River :", "value": "08188500", "class": "II" },
-	{ "text": "Pedernales River : Fredericksburg", "value": "08152900", "class": "II" },
-	{ "text": "Pedernales River : Johnson City", "value": "08153500", "class": "II" },
-	{ "text": "Walnut Creek : Austin", "value": "08158200", "class": "III" },
-	{ "text": "Dolan creek", "value": "08449100", "class": "I" }
-]}
+{
+  "data": [
+    { "text": "Brazos River : Palo Pinto", "value": "08089000", "class": "II-III" },
+    { "text": "Brazos River : Glen Rose", "value": "08091000", "class": "II-III" },
+    { "text": "Brazos River : Waco", "value": "08096500", "class": "II-III" },
+    { "text": "Brazos River : Bryan", "value": "08108700", "class": "II-III" },
+    { "text": "Lampasas River : Kempner", "value": "08103800", "class": "II+" },
+    { "text": "Lampasas River : Ding Dong", "value": "08103940", "class": "II+" },
+    { "text": "Lampasas River : Belton", "value": "08104100", "class": "II+" },
+    { "text": "Barton Springs", "value": "08155500", "class": "N/A" },
+    { "text": "Barton Creek : Above Barton Springs", "value": "08155400", "class": "II" },
+    { "text": "Barton Creek : Loop 360", "value": "08155300", "class": "II-III" },
+    { "text": "Barton Creek : Lost Ck Blvd", "value": "08155240", "class": "II-III" },
+    { "text": "Barton Creek : SH 71", "value": "08155200", "class": "II-III+" },
+    { "text": "Colorado River : Austin", "value": "08158000", "class": "I-II" },
+    { "text": "Colorado River : Bastrop", "value": "08159200", "class": "I-II" },
+    { "text": "Onion Creek : Manchaca", "value": "08158827", "class": "II-III" },
+    { "text": "Onion Creek : Hwy 183", "value": "08159000", "class": "II-III" },
+    { "text": "Onion Creek : Driftwood", "value": "08158700", "class": "II-III" },
+    { "text": "Llano River : Llano", "value": "08151500", "class": "II-III" },
+    { "text": "Llano River : Mason", "value": "08150700", "class": "II-III" },
+    { "text": "Llano River : Junction", "value": "08150000", "class": "I-II" },
+    { "text": "San Saba River : Menard", "value": "08144500", "class": "II" },
+    { "text": "San Saba River : San Saba", "value": "08146000", "class": "II" },
+    { "text": "Guadalupe River : Spring Branch", "value": "08167500", "class": "II" },
+    { "text": "Guadalupe River : Sattler", "value": "08167800", "class": "II" },
+    { "text": "Guadalupe River : Comal New Braunfels", "value": "08168500", "class": "II" },
+    { "text": "Guadalupe River : FM 1117 nr Seguin", "value": "08169792", "class": "II" },
+    { "text": "Guadalupe River : Gonzales", "value": "08169845", "class": "II" },
+    { "text": "San Marcos River : San Marcos", "value": "08170500", "class": "II" },
+    { "text": "San Marcos River : Luling", "value": "08172000", "class": "II" },
+    { "text": "Blanco River : Wimberley", "value": "08171000", "class": "II-III" },
+    { "text": "Blanco River : Kyle", "value": "08171300", "class": "II-III" },
+    { "text": "Nueces River : Montell", "value": "08189998", "class": "II" },
+    { "text": "Nueces River : Laguna", "value": "08190000", "class": "II" },
+    { "text": "Nueces River : Uvalde", "value": "08192000", "class": "II" },
+    { "text": "Nueces River : Cotulla", "value": "08194000", "class": "II" },
+    { "text": "Nueces River : Three Rivers", "value": "08210000", "class": "II" },
+    { "text": "Frio River : Concan", "value": "08195000", "class": "II" },
+    { "text": "Rio Grande : Castolon", "value": "08374550", "class": "II-III" },
+    { "text": "Rio Grande : Big Bend", "value": "08375300", "class": "II-III" },
+    { "text": "Pecos River : Orla", "value": "08412500", "class": "II-III" },
+    { "text": "Pecos River : Girvin", "value": "08446500", "class": "II-III" },
+    { "text": "Medina River : Patterson Road", "value": "0817887350", "class": "I-II" },
+    { "text": "Medina River : Bandera", "value": "08178880", "class": "I-II" },
+    { "text": "San Antonio River :", "value": "08188500", "class": "II" },
+    { "text": "Pedernales River : Fredericksburg", "value": "08152900", "class": "II" },
+    { "text": "Pedernales River : Johnson City", "value": "08153500", "class": "II" },
+    { "text": "Walnut Creek : Austin", "value": "08158200", "class": "III" },
+    { "text": "Dolan creek", "value": "08449100", "class": "I" }
+  ]
+}
